Update weston and backport build fix
The build was failing due to the neatvnc version
in the Fedora repos being newer than the upper limit
listed in the weston build files.
Weston is not affected by the changes in neatvnc
so we can raise the upper version limit for neatvnc.

Backports https://gitlab.freedesktop.org/wayland/weston/-/merge_requests/1534.